+++ src/test/java/supernova/whokie/answer/service/AnswerServiceTest.java
@@ -8,10 +8,10 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.data.domain.*;
 import org.springframework.test.util.ReflectionTestUtils;
 import supernova.whokie.answer.Answer;
-import supernova.whokie.answer.controller.dto.AnswerCommand;
 import supernova.whokie.answer.controller.dto.AnswerModel;
 import supernova.whokie.answer.controller.dto.AnswerResponse;
 import supernova.whokie.answer.repository.AnswerRepository;
+import supernova.whokie.answer.service.dto.AnswerCommand;
 import supernova.whokie.friend.Friend;
 import supernova.whokie.friend.infrastructure.repository.FriendRepository;
 import supernova.whokie.global.dto.PagingResponse;
